помещения контакта
allow filter after selecting помещения контакта;
add link "помещения контакта" to the datatable

+++ app/models/property.rb
@@ -1,5 +1,8 @@
+# encoding: utf-8
 class Property < ActiveRecord::Base
   
+  attr_accessor :old_district  
+
   belongs_to :contact
   
   has_many :property_category_items, dependent: :destroy
@@ -13,10 +16,14 @@ class Property < ActiveRecord::Base
   accepts_nested_attributes_for :property_category_items
     
   after_save :update_local_values
-        
+  
   def item(category)
-    item = category_items.find_by_category_id(Category.find_by_un(category.to_s).id)
-    item.name if item
+    cat = Category.find_by_un(category.to_s)
+    cat = Category.find_by_id(category.to_i) unless cat
+    if cat
+      item = category_items.find_by_category_id(cat.id)
+      item.name if item
+    end
   end
   
   def item_color(category)
@@ -28,7 +35,7 @@ class Property < ActiveRecord::Base
     items = category_items.where(category_id: Category.find_by_un('district').id).pluck(:name)
     items.join(', ') if items.present?
   end
-    
+  
   def self.of_user(user)
     if user.is_main
       all
@@ -53,38 +60,41 @@ class Property < ActiveRecord::Base
     end
     
     Property.where(id: id).update_all(rooms: r, floor: f, floors: fs, address: a, state: s)
-=begin    
+    
     images.each do |image|
       if image.image?
-        path = image.image.path
-        _district = Russian.translit(item(:district).to_s.force_encoding('utf-8'))
-        _landmark_old = Russian.translit(landmark_was.to_s.force_encoding('utf-8'))
-        _landmark = Russian.translit(landmark.to_s.force_encoding('utf-8'))
+        _district_old = Russian.translit((old_district.present? ? old_district : '-').to_s.force_encoding('utf-8'))
+        _district = Russian.translit((item(:district) || '-').to_s.force_encoding('utf-8'))
+        _landmark_old = Russian.translit(landmark_was.to_s.gsub(/[\x00\/\\:\*\?\"<>\|]/, '').force_encoding('utf-8')).gsub(/[\xC2\xBB|\xC2\xAB]/, '').to_s.strip
+        _landmark = Russian.translit(landmark.to_s.gsub(/[\x00\/\\:\*\?\"<>\|]/, '').force_encoding('utf-8')).gsub(/[\xC2\xBB|\xC2\xAB]/, '').to_s.strip
         _kee_old = "#{rooms.to_s[/\d+/]}-#{floor.to_s[/\d+/]}-#{floors.to_s[/\d+/]}"
         _kee = "#{r.to_s[/\d+/]}-#{f.to_s[/\d+/]}-#{fs.to_s[/\d+/]}"
         _price_old = [price1_was, price2_was, price3_was].reject(&:blank?).join(' ')
         _price = [price1, price2, price3].reject(&:blank?).join(' ')
         info_old = "#{_kee_old} #{_landmark_old} #{_price_old}"
         info = "#{_kee} #{_landmark} #{_price}"
-        filename = image.image_file_name
-        old_path = Rails.root.join('public', 'photos', _district, info_old)
+        filename = image.short_name  #  image.image_file_name
+        old_path = Rails.root.join('public', 'photos', _district_old, info_old)
         old_path_with_filename = File.join(old_path, filename)
+        
         new_path = Rails.root.join('public', 'photos', _district, info)
         
         if File.exists?(old_path_with_filename) && old_path_with_filename != File.join(new_path, filename)
           FileUtils.mkdir_p new_path
-          FileUtils.mv old_path_with_filename, File.join(new_path, filename)
+          FileUtils.mv old_path_with_filename, File.join(new_path) # , filename)
+          
           # delete old empty folder
-          if image == images.last
+          if image == images.last && Dir[old_path.to_s + '/*'].empty?
             FileUtils.rm_rf old_path
           end
         end
+        
+        image.save # to update the locally saved image_urls
       end
     end
-=end
   end
   
   def self.update_local_values
     CategoryItem.all.each { |i| i.update_properties }
-  end  
+  end
 end
